Finished Heatmap. Still need to test with vive
I redid the heatmap script and initialized positions on a 9 x 9 grid for a heatmap. Whenever the ray collides with an object, it increases the radius and intensity of the properties for that position. Documented code so that other users can understand it for future editting purposes.

Added some text UIs to let the user know on how to start and stop the heat map.